fsck.f2fs: validate free seg count on zns device
NVMe Zoned Namespace devices can have zone-capacity less than zone-size.
Zone-capacity indicates the number of usable blocks in a zone, if
zone-capacity is less than zone-size, then the segments which start
at/after zone-capacity are considered unusable. Only those segments
which start before the zone-capacity are considered as usable and added
to the free_segment_count and free_segment_bitmap of the kernel.

Allow fsck to find the free_segment_count based on the zone-capacity and
compare with checkpoint values.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>